rename filterData to defaultValues and type them
