+++ web-client/src/shared/components/drivers-search/FilterAccordion/index.tsx
@@ -11,15 +11,16 @@ import FormLabel from '@material-ui/core/FormLabel';
 import FormControl from '@material-ui/core/FormControl';
 import FormGroup from '@material-ui/core/FormGroup';
 import FormControlLabel from '@material-ui/core/FormControlLabel';
+import { DefaultValue } from '../Filters/defaultValues'
 
-interface FilterState {
-    [index: number]: { item: boolean };
-}
+// interface FilterState {
+//     [index: number]: { item: boolean };
+// }
 
 type Props = {
-    title?: string;
-    state: FilterState[];
-    handleChange?: (
+    title: string;
+    state: DefaultValue[];
+    handleChange: (
         event: React.ChangeEvent<HTMLInputElement>,
         checked: boolean,
     ) => void;
